# cmd/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "User friendly tax deduction system",
        "title": "Info Tax",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:5000",
    "basePath": "/",
    "paths": {
...
        "/infotax/user_login_detail": {
            "post": {
                "description": "Create login user details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Login Detail"
                ],
                "summary": "Create login user details",
                "operationId": "create_user_login_detail",
                "parameters": [
                    {
                        "description": "User login details",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CreateUserDetail"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SaveResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "415": {
                        "description": "Unsupported Media Type",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/infotax/user_login_detail/": {
            "get": {
                "description": "Get all login user detail",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Login Detail"
                ],
                "summary": "Get all login user detail",
                "operationId": "get_all_user_login_detail",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/UserLoginDetail"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
...
        "/infotax/user_login_detail/role/{emp_id}": {
            "get": {
                "description": "Get role by employee id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Login Detail"
                ],
                "summary": "Get login user role based on employee id",
                "operationId": "get_user_login_detail_role",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Employee ID",
                        "name": "emp_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/UserLoginRole"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/infotax/user_login_detail/search": {
            "get": {
                "description": "Search login user detail based on search filter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Login Detail"
                ],
                "summary": "Search login user detail based on search filter",
                "operationId": "search_user_login_detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "EmployeeID filter",
                        "name": "employee_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "DomainName filter",
                        "name": "domain_name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "EmailID filter",
                        "name": "email_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Password filter",
                        "name": "password",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "UUID filter",
                        "name": "uuid",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "IsSignedup filter",
                        "name": "isSignedup",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Role filter",
                        "name": "role",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/UserLoginDetail"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/infotax/user_login_detail/upload": {
            "post": {
                "description": "Insert bulk of users login details present in input csv file",
                "consumes": [
                    "application/octet-stream"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Login Detail"
                ],
                "summary": "Insert bulk of users login details present in input csv file",
                "operationId": "upload_users_login_detail",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Users login details form",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SaveResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/infotax/user_login_detail/{emp_id}": {
            "get": {
                "description": "Get detail by employee id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Login Detail"
                ],
                "summary": "Get login user detail based on employee id",
                "operationId": "get_user_login_detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Employee ID",
                        "name": "emp_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/UserLoginDetail"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "put": {
                "description": "Update login user details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Login Detail"
                ],
                "summary": "Update login user details",
                "operationId": "update_user_login_detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Employee ID",
                        "name": "emp_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update user login details",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UpdateUserDetail"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/UpdateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "415": {
                        "description": "Unsupported Media Type",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete login user details by employee id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Login Detail"
                ],
                "summary": "Delete login user details based on employee id",
                "operationId": "delete_user_login_detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Employee ID",
                        "name": "emp_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/DeleteResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "CreateUserDetail": {
            "type": "object",
            "required": [
                "domain_name",
                "email_id",
                "employee_id",
                "role"
            ],
            "properties": {
                "domain_name": {
                    "type": "string"
                },
                "email_id": {
                    "type": "string"
                },
                "employee_id": {
                    "type": "string"
                },
                "role": {
                    "type": "string",
                    "enum": [
                        "Payroll_User",
                        "Accounts_Team",
                        "Payroll_Team"
                    ]
                }
            }
        },
        "DeleteResponse": {
            "type": "object",
            "properties": {
                "isDeleted": {
                    "type": "string"
                }
            }
        },
        "Error": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "SaveResponse": {
            "type": "object",
            "properties": {
                "isSaved": {
                    "type": "string"
                }
            }
        },
        "UpdateResponse": {
            "type": "object",
            "properties": {
                "isUpdated": {
                    "type": "string"
                }
            }
        },
        "UpdateUserDetail": {
            "type": "object",
            "required": [
                "domain_name"
            ],
            "properties": {
                "domain_name": {
                    "type": "string"
                },
                "email_id": {
                    "type": "string"
                },
                "role": {
                    "type": "string",
                    "enum": [
                        "Payroll_User",
                        "Accounts_Team",
                        "Payroll_Team"
                    ]
                }
            }
        },
        "UserLoginDetail": {
            "type": "object",
            "properties": {
                "domain_name": {
                    "type": "string"
                },
                "email_id": {
                    "type": "string"
                },
                "employee_id": {
                    "type": "string"
                },
                "isSignedup": {
                    "type": "integer"
                },
                "password": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "UserLoginRole": {
            "type": "object",
            "properties": {
                "role": {
                    "type": "string"
                }
            }
        },
...
    }
}